Update custom video renderer
