+++ common/gradle.properties
@@ -1,0 +1,1 @@
+android.useAndroidX=true
