+++ exampleAdvancedCameraCapturer/build.gradle
@@ -39,6 +39,7 @@ android {
 dependencies {
     testImplementation 'junit:junit:4.12'
 
+    implementation project(":common")
     implementation "com.twilio:video-android:${versions.videoAndroid}"
     implementation "com.android.support:appcompat-v7:${versions.supportLibrary}"
     implementation "com.android.support:design:${versions.supportLibrary}"
