+++ exampleCustomVideoRenderer/src/main/java/com/twilio/video/examples/customrenderer/SnapshotVideoRenderer.kt
@@ -1,0 +1,40 @@
+package com.twilio.video.examples.customrenderer
+
+import android.os.Handler
+import android.os.Looper
+import android.widget.ImageView
+import com.twilio.video.examples.common.toBitmap
+import tvi.webrtc.VideoFrame
+import tvi.webrtc.VideoSink
+import java.util.concurrent.atomic.AtomicBoolean
+
+/**
+ * SnapshotVideoRenderer demonstrates how to implement a custom [tvi.webrtc.VideoSink]. Caches the
+ * last frame rendered and will update the provided image view any time [.takeSnapshot] is
+ * invoked.
+ */
+class SnapshotVideoRenderer(private val imageView: ImageView) : VideoSink {
+    private val snapshotRequsted =
+        AtomicBoolean(false)
+    private val handler = Handler(Looper.getMainLooper())
+
+    override fun onFrame(videoFrame: VideoFrame) {
+        videoFrame.retain()
+        if (snapshotRequsted.compareAndSet(true, false)) {
+            val bitmap = videoFrame.toBitmap()
+            handler.post {
+                imageView.setImageBitmap(bitmap)
+                videoFrame.release()
+            }
+        } else {
+            videoFrame.release()
+        }
+    }
+
+    /**
+     * Request a snapshot on the rendering thread.
+     */
+    fun takeSnapshot() {
+        snapshotRequsted.set(true)
+    }
+}
